requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-crispy-forms==1.8.1
django-formtools==2.2
django-import-export==2.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.5
django-simple-captcha==0.5.12
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.4
httplib2==0.15.0
idna==2.8
ipython==7.10.2
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.2
MarkupPy==1.14
mysqlclient==1.4.6
odfpy==1.4.0
openpyxl==3.0.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==3.0.2
ptyprocess==0.6.0
Pygments==2.5.2
pynvim==0.4.0
pytz==2019.3
PyYAML==5.2
redis==3.3.11
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
